implement a query that returns the cheapest 3 properties in ranges between 40 and 90 squaremetres

+++ statistics/web/query/behavior/cheapest.ex
@@ -1,5 +1,5 @@
 defmodule Statistics.Query.Behaviour.Cheapest do
   alias Statistics.Property
 
-  @callback execute(integer, float, float) :: [Property]
+  @callback execute(integer, integer, integer) :: [Property]
 end
